Add first test for integer premetive
