--- lisp_interpreter/tests/test_lisp_interpreter.py
@@ -1,13 +1,0 @@
-import sys
-sys.path.append("..")
-
-import unittest
-
-
-class TestFizzBuzz(unittest.TestCase):
-    def test_run(self):
-        self.assertTrue(True)
-
-
-if __name__ == '__main__':
-    unittest.main()
